chore: Refactor database-related interfaces and classes
