+++ src/ProjectOrigin.Chronicler.Server/Repositories/ChroniclerRepository.cs
@@ -2,7 +2,7 @@ using System.Data;
 using System.Threading.Tasks;
 using Dapper;
 using ProjectOrigin.Chronicler.Server.Models;
-using ProjectOrigin.ServiceCommon.DataPersistence;
+using ProjectOrigin.ServiceCommon.Database;
 using ProjectOrigin.WalletSystem.Server.Repositories;
 public class ChroniclerRepository : AbstractRepository, IChroniclerRepository
 {
