--- src/ProjectOrigin.ServiceCommon/DataPersistence/IDatebaseUpgrader.cs
@@ -1,7 +1,0 @@
-namespace ProjectOrigin.ServiceCommon.DataPersistence;
-
-public interface IDatebaseUpgrader
-{
-    Task Upgrade();
-    Task<bool> IsUpgradeRequired();
-}
