--- src/ProjectOrigin.ServiceCommon/DataPersistence/IDbConnectionFactory.cs
@@ -1,8 +1,0 @@
-using System.Data;
-
-namespace ProjectOrigin.ServiceCommon.DataPersistence;
-
-public interface IDbConnectionFactory
-{
-    IDbConnection CreateConnection();
-}
