+++ src/ProjectOrigin.ServiceCommon/Database/IDatabaseConnectionFactory.cs
@@ -1,0 +1,8 @@
+using System.Data;
+
+namespace ProjectOrigin.ServiceCommon.Database;
+
+public interface IDatabaseConnectionFactory
+{
+    IDbConnection CreateConnection();
+}
