+++ src/ProjectOrigin.ServiceCommon/Database/IDatabaseUpgrader.cs
@@ -1,0 +1,7 @@
+namespace ProjectOrigin.ServiceCommon.Database;
+
+public interface IDatabaseUpgrader
+{
+    Task Upgrade();
+    Task<bool> IsUpgradeRequired();
+}
