+++ src/ProjectOrigin.ServiceCommon/Database/Postgres/PostgresConfigurationBuilder.cs
@@ -2,9 +2,9 @@ using System.Data;
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace ProjectOrigin.ServiceCommon.DataPersistence.Postgres;
+namespace ProjectOrigin.ServiceCommon.Database.Postgres;
 
-internal class PostgresConfigurationBuilder : IDataPersistenceConfigurationBuilder
+internal class PostgresConfigurationBuilder : IDatabaseConfigurationBuilder
 {
     public IEnumerable<Assembly> DatabaseScriptsAssemblies = new List<Assembly>();
     public IDictionary<Type, ObjectFactory> RepositoryFactories = new Dictionary<Type, ObjectFactory>();
